+++ src/historyView/itemLogProvider.ts
@@ -1,3 +1,4 @@
+import { registerResources } from "../lifecycle";
 import * as path from "path";
 import {
   commands,
@@ -47,29 +48,39 @@ export class ItemLogProvider
   private disposed = false;
 
   constructor(private sourceControlManager: SourceControlManager) {
-    this._dispose.push(
-      window.onDidChangeActiveTextEditor(this.editorChanged, this),
-      window.registerTreeDataProvider("itemlog", this),
-      commands.registerCommand(
-        "svn.itemlog.copymsg",
-        async (item: ILogTreeItem) => copyCommitToClipboard("msg", item)
-      ),
-      commands.registerCommand(
-        "svn.itemlog.copyrevision",
-        async (item: ILogTreeItem) => copyCommitToClipboard("revision", item)
-      ),
-      commands.registerCommand(
-        "svn.itemlog.openFileRemote",
-        this.openFileRemoteCmd,
-        this
-      ),
-      commands.registerCommand("svn.itemlog.openDiff", this.openDiffCmd, this),
-      commands.registerCommand(
-        "svn.itemlog.openDiffBase",
-        this.openDiffBaseCmd,
-        this
-      ),
-      commands.registerCommand("svn.itemlog.refresh", this.refresh, this)
+    registerResources(
+      this._dispose,
+      () => window.onDidChangeActiveTextEditor(this.editorChanged, this),
+      () => window.registerTreeDataProvider("itemlog", this),
+      () =>
+        commands.registerCommand(
+          "svn.itemlog.copymsg",
+          async (item: ILogTreeItem) => copyCommitToClipboard("msg", item)
+        ),
+      () =>
+        commands.registerCommand(
+          "svn.itemlog.copyrevision",
+          async (item: ILogTreeItem) => copyCommitToClipboard("revision", item)
+        ),
+      () =>
+        commands.registerCommand(
+          "svn.itemlog.openFileRemote",
+          this.openFileRemoteCmd,
+          this
+        ),
+      () =>
+        commands.registerCommand(
+          "svn.itemlog.openDiff",
+          this.openDiffCmd,
+          this
+        ),
+      () =>
+        commands.registerCommand(
+          "svn.itemlog.openDiffBase",
+          this.openDiffBaseCmd,
+          this
+        ),
+      () => commands.registerCommand("svn.itemlog.refresh", this.refresh, this)
     );
     this.refresh();
   }